backend/requirements/base.txt
Django~=5.0.0
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0
python-decouple~=3.8
dj-database-url~=2.1.0
psycopg2-binary~=2.9.9
django-filter~=23.5
drf-spectacular~=0.26.5
Django~=5.0.0
celery~=5.3.6
redis~=5.0.1
django-leaflet~=0.32.0
Pillow~=10.0.1 # Often needed for image uploads even if not directly specified yet